// packages/remote-agent-daemon/src/codex/generated/v2/ToolRequestUserInputResponse.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ToolRequestUserInputAnswer } from "./ToolRequestUserInputAnswer.js"

/**
 * EXPERIMENTAL. Response payload mapping question ids to answers.
 */
export type ToolRequestUserInputResponse = {
  answers: { [key in string]?: ToolRequestUserInputAnswer }
}